Fix nt2 installation
* avoid deletion of inexistant folder

+++ setup.py
@@ -51,7 +51,8 @@ class BuildWithPly(build):
         for d in ('nt2', 'boost'):
             src = os.path.join(nt2_build_dir, 'include', d)
             target = os.path.join(self.build_lib, 'pythran', d)
-            shutil.rmtree(target)
+            if os.path.exists(target):
+                shutil.rmtree(target)
             shutil.copytree(src, target)
 
     def run(self, *args, **kwargs):
